slotを使ったコンポーネント
HTMLを送った
