--- src/LikeHeader.vue
@@ -1,3 +1,0 @@
-<template>
-  <h1>いいねボタン表示</h1>
-</template>
