--- src/LikeNumber.vue
@@ -1,27 +1,0 @@
-<template>
-  <div>
-    <p>いいね({{ number }})</p>
-    <button v-on:click="increment">+1</button>
-  </div>
-</template>
-
-<script>
-export default {
-  data: function() {
-    return {
-      number: 5,
-    };
-  },
-  methods: {
-    increment: function() {
-      this.number += 1;
-    }
-  },
-};
-</script>
-
-<style scoped>
-  div {
-    border: 5px solid red;
-  }
-</style>
