+++ src/components/LikeHeader.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <slot></slot>
+    <slot></slot>
+    <p>{{ headerText }}</p>
+  </div>
+</template>
+
+<script>
+export default {
+  props: ["headerText"]
+};
+</script>
